modified entity article

+++ src/main/java/com/belvinard/gestiondestock/repositories/LigneCommandeClientRepository.java
@@ -10,5 +10,5 @@ public interface LigneCommandeClientRepository extends JpaRepository<LigneComman
 
   List<LigneCommandeClient> findAllByCommandeClientId(Integer id);
 
-  List<LigneCommandeClient> findAllByArticleId(Integer id);
+  List<LigneCommandeClient> findAllByArticleId(Long id);
 }
